+++ mound/__init__.py
@@ -3,7 +3,7 @@
 from mound.pitches import Batter, PitchCollection, Pitcher
 from mound.players import AmbiguousPlayerError, Player, PlayerNotFoundError
 
-__version__ = "0.7.0"
+__version__ = "0.7.1"
 
 __all__ = [
     "Pitcher",